Move date related util functions to separate file
