+++ source/frontend/src/utils/index.ts
@@ -1,6 +1,7 @@
 export * from './axiosUtils';
 export * from './columnUtils';
 export * from './convertUtils';
+export * from './dateUtils';
 export * from './mapLookupCode';
 export * from './numberFormatUtils';
 export * from './projectUtils';
