+++ source/frontend/src/utils/utils.ts
@@ -1,7 +1,6 @@
 import { AxiosError, AxiosResponse } from 'axios';
 import { FormikProps, getIn } from 'formik';
 import { isEmpty, isNull, isUndefined, keys, lowerFirst, startCase } from 'lodash';
-import moment, { Moment } from 'moment-timezone';
 import { hideLoading, showLoading } from 'react-redux-loading-bar';
 
 import { SelectOption } from '@/components/common/form';
@@ -157,32 +156,6 @@ export const resolveSortCriteriaFromUrl = (input: string[]): TableSort<any> | {}
 
     return { ...acc, [lowerFirst(fields[0])]: fields[1] };
   }, {});
-};
-
-/**
- * get the fiscal year (ending in) based on the current date.
- */
-export const getCurrentFiscalYear = (): number => {
-  const now = moment();
-  return now.month() >= 4 ? now.add(1, 'years').year() : now.year();
-};
-
-export const prettyFormatDate = (date?: string | Date | Moment | null) => {
-  return !!date ? moment(date).format('MMM D, YYYY') : '';
-};
-
-export const prettyFormatDateTime = (date?: string | Date | Moment | null) => {
-  return !!date ? moment.utc(date).local().format('MMM D, YYYY hh:mm a') : '';
-};
-
-/**
- * Format the passed string date assuming the date was recorded in UTC (as is the case with the pims API server).
- * Returns a date formatted for display in the current time zone of the user.
- * @param date utc date/time string.
- */
-export const formatApiDateTime = (date?: string | Date | Moment) => {
-  if (typeof date === 'string') return moment.utc(date).local().format('YYYY-MM-DD hh:mm a');
-  return !!date ? moment.utc(date).local().format('YYYY-MM-DD hh:mm a') : '';
 };
 
 /**
